Move Cavern class to upper level
